More rendering fixes from R2D2_C3PO
The area for tourism=attraction is not rendered when a highway=*
attribute is present or when area=no is set. Previously this had lead to
a problem with a bridge which was not a completely straight line.

For tourism=attraction and amenity=place_of_worship or historic=ruins
only one name is rendered and not two times in different colors.

When building=yes and amenity=fuel are set the text is correctly offset.