[renderd] support alternative map projections other than web mercator
Remove hard coded references to the spherical mercator projection and
replace them with something a little more generic. It is still necessary
to add code to support new projections, but it is limited to a few configuration numbers

This initial commit supports spherical mercator and plate carree.

This commit also makes the tile size configurable. The default remains at 256 px.